added basic admin functionalities

+++ account/views.py
@@ -11,17 +11,18 @@ from django.contrib.sites.shortcuts import get_current_site
 from django.core.exceptions import ObjectDoesNotExist
 from django.core.mail import send_mail
 from django.db.models.query_utils import Q
-from django.http import Http404
+from django.http import Http404, HttpResponse
 from django.shortcuts import render, redirect, get_object_or_404
 from django.template import loader
 from django.urls import reverse_lazy
 from django.utils.encoding import force_bytes
 from django.utils.http import urlsafe_base64_encode
-from django.views.generic import FormView
+from django.views.generic import FormView, TemplateView
 
-from .forms import (RegistrationForm, LoginForm, CustomerProfileForm, PharmacyProfileForm, UpdateUserForm,
+from .forms import (RegistrationForm, LoginForm, UpdateUserForm,
+                    CustomerProfileForm, PharmacyProfileForm, PharmaAdminProfileForm,
                     ConfirmUserPasswordResetForm, RequestUserPasswordResetForm, UserPasswordChangeForm)
-from .models import Customer, Pharmacy
+from .models import Customer, Pharmacy, PharmaAdmin
 
 from store.views import get_order_count, get_cart, get_cart_count
 
@@ -32,45 +33,48 @@ def index(request):
 
 
 def register_view(request, user_label):
-    if not request.user.is_authenticated:
-        form = RegistrationForm()
-        if user_label == 'customer':
-            profile_form = CustomerProfileForm()
-        elif user_label == 'pharmacy':
-            profile_form = PharmacyProfileForm()
-        else:
-            profile_form = None
-
-        if request.method == 'POST':
-            form = RegistrationForm(request.POST)
-            if user_label == 'customer':
-                profile_form = CustomerProfileForm(request.POST)
-            elif user_label == 'pharmacy':
-                profile_form = PharmacyProfileForm(request.POST)
-
-            if form.is_valid() and profile_form.is_valid():
-                user = form.save()
-                profile = profile_form.save(commit=False)
-                profile.user = user
-                profile.save()
-                messages.success(request, f"Account created for {form.cleaned_data.get('username')}!")
-                return redirect('account:login', user_label=user_label)
-
-        if user_label == 'customer':
-            form_header = 'Create Customer Account'
-        elif user_label == 'pharmacy':
-            form_header = 'Create Pharmacy Account'
-        else:
+    form = RegistrationForm()
+    if user_label == 'customer':
+        profile_form = CustomerProfileForm()
+    elif user_label == 'pharmacy':
+        # make sure that only pharma-admins can create pharmacy account
+        try:
+            get_object_or_404(PharmaAdmin, user=request.user)
+        except TypeError:
             raise Http404('Page not found')
-
-        if profile_form:
-            context = {'form': form, 'profile_form': profile_form,
-                       'form_header': form_header, 'submit_msg': 'Sign Up'}
-
-        return render(request, 'account/user_register.html', context=context)
+        profile_form = PharmacyProfileForm()
     else:
-        messages.warning(request, 'You\'re already logged in. Please logout first.')
-        return redirect('store:home')
+        profile_form = None
+
+    if request.method == 'POST':
+        form = RegistrationForm(request.POST)
+        if user_label == 'customer':
+            profile_form = CustomerProfileForm(request.POST)
+        elif user_label == 'pharmacy':
+            profile_form = PharmacyProfileForm(request.POST)
+
+        if form.is_valid() and profile_form.is_valid():
+            user = form.save()
+            profile = profile_form.save(commit=False)
+            profile.user = user
+            profile.save()
+            messages.success(request, f"Account created for {form.cleaned_data.get('username')}!")
+            if user_label == 'pharmacy':
+                return redirect('store:pharma_admin_home')
+            return redirect('account:login', user_label=user_label)
+
+    if user_label == 'customer':
+        form_header = 'Create Customer Account'
+    elif user_label == 'pharmacy':
+        form_header = 'Create Pharmacy Account'
+    else:
+        raise Http404('Page not found')
+
+    if profile_form:
+        context = {'form': form, 'profile_form': profile_form,
+                   'form_header': form_header, 'submit_msg': 'Sign Up'}
+
+    return render(request, 'account/user_register.html', context=context)
 
 
 def login_view(request, user_label):
@@ -87,15 +91,17 @@ def login_view(request, user_label):
                 if user:
                     try:
                         if user_label == 'customer' and Customer.objects.get(user=user):
-                            pass
+                            redirect_to = 'store:home'
                         elif user_label == 'pharmacy' and Pharmacy.objects.get(user=user):
-                            pass
+                            redirect_to = 'store:home'
+                        elif user_label == 'pharma_admin' and PharmaAdmin.objects.get(user=user):
+                            redirect_to = 'store:pharma_admin_home'
                     except ObjectDoesNotExist:
                         messages.error(request, 'Username or password is incorrect.')
                     else:
                         login(request, user)
                         messages.success(request, 'Login successful!')
-                        return redirect('store:home')
+                        return redirect(redirect_to)
                 else:
                     messages.error(request, 'Username or password is incorrect')
 
@@ -107,6 +113,8 @@ def login_view(request, user_label):
             context['form_header'] = 'Customer Login'
         elif user_label == 'pharmacy':
             context['form_header'] = 'Pharmacy Login'
+        elif user_label == 'pharma_admin':
+            context['form_header'] = 'Administrator Login'
         else:
             raise Http404("Page not found")
 
@@ -127,43 +135,56 @@ def logout_view(request):
     return redirect('store:home')
 
 
-def update_form(form, profile_form):
+def update_user(form, profile_form):
     user = form.save()
     profile = profile_form.save(commit=False)
     profile.user = user
     profile.save()
 
 
+# make sure that only account owners can view their profile page
 def proflie_view(request, user_label, fk):
     if request.user.is_authenticated:
         user_obj = get_object_or_404(User, id=fk)
-        form = UpdateUserForm(instance=user_obj)
+        if user_obj != request.user:
+            raise Http404('Page not found')
+        form = UpdateUserForm(instance=user_obj, initial={'id': fk})
         if user_label == 'customer':
-            customer_obj = Customer.objects.get(user=user_obj)
+            customer_obj = get_object_or_404(Customer, user=user_obj)
             profile_form = CustomerProfileForm(instance=customer_obj)
         elif user_label == 'pharmacy':
-            pharmacy_obj = Pharmacy.objects.get(user=user_obj)
+            pharmacy_obj = get_object_or_404(Pharmacy, user=user_obj)
             profile_form = PharmacyProfileForm(instance=pharmacy_obj)
+        elif user_label == 'pharma_admin':
+            pharma_admin_obj = get_object_or_404(PharmaAdmin, user=user_obj)
+            profile_form = PharmaAdminProfileForm(instance=pharma_admin_obj)
         else:
             profile_form = None
 
         if request.method == 'POST':
-            form = UpdateUserForm(request.POST, instance=user_obj)
+            form = UpdateUserForm(request.POST, instance=user_obj, initial={'id': fk})
             if user_label == 'customer':
                 profile_form = CustomerProfileForm(request.POST, instance=customer_obj)
             elif user_label == 'pharmacy':
                 profile_form = PharmacyProfileForm(request.POST, instance=pharmacy_obj)
+            elif user_label == 'pharma_admin':
+                profile_form = PharmaAdminProfileForm(request.POST, instance=pharma_admin_obj)
 
             if form.is_valid() and profile_form.is_valid():
                 if form.has_changed() or profile_form.has_changed():
-                    update_form(form, profile_form)
+                    update_user(form, profile_form)
                     messages.success(request, f"Account updated for {form.cleaned_data.get('username')}!")
+
+                if user_label == 'pharma_admin':
+                    return redirect('store:pharma_admin_home')
                 return redirect('store:home')
 
         if user_label == 'customer':
             form_header = 'Customer Profile'
         elif user_label == 'pharmacy':
             form_header = 'Pharmacy Profile'
+        elif user_label == 'pharma_admin':
+            form_header = 'Administrator Profile'
         else:
             raise Http404('Page not found')
 
@@ -181,11 +202,15 @@ def proflie_view(request, user_label, fk):
 
 def delete_view(request, pk):
     if request.user.is_authenticated:
-        user_obj = get_object_or_404(User, id=pk)
-        if request.method == "POST":
-            user_obj.delete()
-            messages.success(request, 'Account Deleted')
-            return redirect('store:home')
+        # make sure that only a user can delete it's account from this view
+        if request.user.id == pk:
+            user_obj = get_object_or_404(User, id=pk)
+            if request.method == "POST":
+                user_obj.delete()
+                messages.success(request, 'Account Deleted')
+                return redirect('store:home')
+        else:
+            raise Http404('Page not found')
         return render(request, 'account/user_profile.html')
     else:
         messages.warning(request, 'No logged in user found.')
@@ -274,5 +299,55 @@ class UserPasswordChangeView(PasswordChangeView):
                             kwargs={'user_label': user_label, 'fk': self.request.user.id})
 
 
-def admin_homepage_view(request):
-    return render(request, 'account/admin/index.html')
+def admin_manage_users_view(request, manage, user_label):
+    # make sure that only admin can access this functionality
+    try:
+        get_object_or_404(PharmaAdmin, user=request.user)
+    except TypeError:
+        raise Http404('Page not found')
+
+    if user_label == 'pharmacy' and manage in ['delete', 'disable', 'enable']:
+        pass
+    elif user_label == 'customer' and manage in ['disable', 'enable']:
+        pass
+    else:
+        raise Http404('Page not found')
+
+    if request.method == 'POST':
+        username = request.POST.get('username', None)
+        if not username:
+            messages.error(request, 'Field cannot be empty.')
+        else:
+            try:
+                user_obj = User.objects.get(username=username)
+                target_user = None
+                if user_label == 'pharmacy':
+                    target_user = Pharmacy.objects.filter(user=user_obj).first()
+                elif user_label == 'customer':
+                    target_user = Customer.objects.filter(user=user_obj).first()
+
+                if target_user:
+                    if manage == 'delete':
+                        user_obj.delete()
+                        success_msg = f'{user_obj} account deleted'
+                    elif manage == 'disable':
+                        target_user.disabled = True
+                        target_user.save()
+                        success_msg = f'{user_obj} account disabled'
+                    elif manage == 'enable':
+                        target_user.disabled = False
+                        target_user.save()
+                        success_msg = f'{user_obj} account enabled'
+                else:
+                    raise ObjectDoesNotExist
+            except ObjectDoesNotExist:
+                messages.error(request, 'Account not found.')
+            else:
+                messages.success(request, success_msg)
+                return redirect('store:pharma_admin_home')
+
+    context = {
+        'msg': manage,
+        'user_label': user_label
+    }
+    return render(request, 'account/admin_user_manage.html', context=context)
